requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
google-generativeai>=0.3.2
langchain>=0.1.0
langchain-google-genai>=0.0.8
python-multipart>=0.0.6
python-dotenv>=1.0.0
supabase>=2.0.0
pillow>=10.1.0 